LayoutTests/platform/mac/media/video-controls-captions-trackmenu-localized-expected.txt

Test that captions and subtitles menu items are localized.

EVENT(canplaythrough)

*** Set the user language preference.
RUN(internals.setUserPreferredLanguages(['en']))
Heading should be labelled 'Subtitles'
EXPECTED (trackListSection.textContent == 'Subtitles') OK

First item in captions menu should be labelled 'Off'
EXPECTED (item.textContent == 'Off') OK

Second item in captions menu should be labelled 'English CC'
EXPECTED (item.textContent == 'English CC'), OBSERVED 'Unknown SDH' FAIL

Third item in captions menu should be labelled 'Unknown SDH'
EXPECTED (item.textContent == 'Unknown SDH'), OBSERVED 'Unknown' FAIL

Fourth item in captions menu should be labelled 'Unknown'
TypeError: 'undefined' is not an object (evaluating 'item.textContent')
END OF TEST
